refactor: merge model files
