--- src/main/kotlin/model/EnumEntry.kt
@@ -1,10 +1,0 @@
-package model
-
-/**
- * Enum entry definition data.
- */
-data class EnumEntry(
-    val name: String,
-    val description: String,
-    val value: Int,
-)
